Refactor airnode examples after introducing template requests
