+++ packages/airnode-examples/integrations/coingecko-post-processing/request-utils.ts
@@ -1,11 +1,34 @@
 import { encode } from '@api3/airnode-abi';
-import { cliPrint, getDeployedContract, readIntegrationInfo } from '../../src';
+import { deriveAirnodeXpub, deriveSponsorWalletAddress } from '@api3/airnode-admin';
+import { ethers } from 'ethers';
+import { cliPrint, getAirnodeWallet, getDeployedContract, readConfig, readIntegrationInfo } from '../../src';
 
-export const getEncodedParameters = () => {
-  return encode([
+const getEncodedParameters = () =>
+  encode([
     { name: 'coinIds', type: 'string32', value: 'bitcoin,ethereum' },
     { name: 'vsCurrency', type: 'string32', value: 'usd' },
   ]);
+
+export const makeRequest = async () => {
+  const integrationInfo = readIntegrationInfo();
+  const requester = await getDeployedContract(`contracts/${integrationInfo.integration}/Requester.sol`);
+  const airnodeWallet = getAirnodeWallet();
+  const sponsor = ethers.Wallet.fromMnemonic(integrationInfo.mnemonic);
+  const endpointId = readConfig().triggers.rrp[0].endpointId;
+  const sponsorWalletAddress = await deriveSponsorWalletAddress(
+    deriveAirnodeXpub(airnodeWallet.mnemonic.phrase),
+    airnodeWallet.address,
+    sponsor.address
+  );
+
+  // Trigger the Airnode request
+  return requester.makeRequest(
+    airnodeWallet.address,
+    endpointId,
+    sponsor.address,
+    sponsorWalletAddress,
+    getEncodedParameters()
+  );
 };
 
 export const printResponse = async (requestId: string) => {
