+++ packages/airnode-examples/integrations/failing-example/request-utils.ts
@@ -1,11 +1,33 @@
 import { encode } from '@api3/airnode-abi';
-import { cliPrint, getDeployedContract, readIntegrationInfo } from '../../src';
+import { deriveAirnodeXpub, deriveSponsorWalletAddress } from '@api3/airnode-admin';
+import { ethers } from 'ethers';
+import { cliPrint, getAirnodeWallet, getDeployedContract, readConfig, readIntegrationInfo } from '../../src';
 
 const coinLabel = 'Ethereum';
 const coinId = coinLabel.toLowerCase();
 
-export const getEncodedParameters = () => {
-  return encode([{ name: 'coinId', type: 'string32', value: coinId }]);
+const getEncodedParameters = () => encode([{ name: 'coinId', type: 'string32', value: coinId }]);
+
+export const makeRequest = async () => {
+  const integrationInfo = readIntegrationInfo();
+  const requester = await getDeployedContract(`contracts/${integrationInfo.integration}/Requester.sol`);
+  const airnodeWallet = getAirnodeWallet();
+  const sponsor = ethers.Wallet.fromMnemonic(integrationInfo.mnemonic);
+  const endpointId = readConfig().triggers.rrp[0].endpointId;
+  const sponsorWalletAddress = await deriveSponsorWalletAddress(
+    deriveAirnodeXpub(airnodeWallet.mnemonic.phrase),
+    airnodeWallet.address,
+    sponsor.address
+  );
+
+  // Trigger the Airnode request
+  return requester.makeRequest(
+    airnodeWallet.address,
+    endpointId,
+    sponsor.address,
+    sponsorWalletAddress,
+    getEncodedParameters()
+  );
 };
 
 export const printResponse = async (requestId: string) => {
