+++ packages/airnode-examples/scripts/make-request.ts
@@ -1,12 +1,9 @@
 import { ethers } from 'ethers';
-import { deriveAirnodeXpub, deriveSponsorWalletAddress } from '@api3/airnode-admin';
 import {
   getDeployedContract,
   getProvider,
   readIntegrationInfo,
   runAndHandleErrors,
-  getAirnodeWallet,
-  readConfig,
   cliPrint,
   setMaxPromiseTimeout,
 } from '../src';
@@ -34,33 +31,12 @@ const waitForFulfillment = async (requestId: string) => {
 
 const makeRequest = async (): Promise<string> => {
   const integrationInfo = readIntegrationInfo();
-  const requester = await getDeployedContract(`contracts/${integrationInfo.integration}/Requester.sol`);
   const airnodeRrp = await getDeployedContract('@api3/airnode-protocol/contracts/rrp/AirnodeRrpV0.sol');
-  const airnodeWallet = getAirnodeWallet();
-  const sponsor = ethers.Wallet.fromMnemonic(integrationInfo.mnemonic);
-  // NOTE: The request is always made to the first endpoint listed in the "triggers.rrp" inside config.json
-  const endpointId = readConfig().triggers.rrp[0].endpointId;
-  // NOTE: When doing this manually, you can use the 'derive-sponsor-wallet-address' and 'derive-airnode-xpub' commands
-  // from the admin CLI package
-  const sponsorWalletAddress = await deriveSponsorWalletAddress(
-    deriveAirnodeXpub(airnodeWallet.mnemonic.phrase),
-    airnodeWallet.address,
-    sponsor.address
-  );
 
-  // Import the function to get encoded parameters for the Airnode. See the respective "request-utils.ts" for details.
-  const { getEncodedParameters } = await import(`../integrations/${integrationInfo.integration}/request-utils.ts`);
-  // Trigger the Airnode request
-
-  const receipt = await requester.makeRequest(
-    // TODO: Move "makeRequest" into request-utils of each integration + update docs
-    '0x02834eb43d56133982b7d6e5aa8b466c7ea4ba0fadf697698c1fee0996bba0fc',
-    // airnodeWallet.address,
-    // endpointId,
-    sponsor.address,
-    sponsorWalletAddress,
-    await getEncodedParameters()
-  );
+  // Import the "makeRequest" which triggers the Airnode request.
+  // See the "request-utils.ts" of the specific integration for details.
+  const { makeRequest } = await import(`../integrations/${integrationInfo.integration}/request-utils.ts`);
+  const receipt = await makeRequest();
 
   // Wait until the transaction is mined
   return new Promise((resolve) =>
